Clean up build without xalan or openssl
git-svn-id: https://svn.apache.org/repos/asf/xml/security/trunk@351000 13f79535-47bb-0310-9956-ffa450edef68

+++ c/src/samples/simpleDecrypt.cpp
@@ -126,7 +126,7 @@ int main (int argc, char **argv) {
 
 	// Create an input source
 
-	MemBufInputSource* memIS = new MemBufInputSource ((const XMLByte*) letter, strlen(letter), "XSECMem");
+	MemBufInputSource* memIS = new MemBufInputSource ((const XMLByte*) letter, (unsigned int) strlen(letter), "XSECMem");
 
 	int errorCount = 0;
